// grpc-service/src/protos/pillow/DataDto.ts
// Original file: src/protos/pillow.proto


export interface DataDto {
  'snoringRange'?: (string);
  'respirationRate'?: (string);
  'bodyTemperature'?: (string);
  'limbMovement'?: (string);
  'bloodOxygen'?: (string);
  'rem'?: (string);
  'hoursSleeping'?: (string);
  'heartRate'?: (string);
  'stresState'?: (string);
}

export interface DataDto__Output {
  'snoringRange': (string);
  'respirationRate': (string);
  'bodyTemperature': (string);
  'limbMovement': (string);
  'bloodOxygen': (string);
  'rem': (string);
  'hoursSleeping': (string);
  'heartRate': (string);
  'stresState': (string);
}