// grpc-service/src/protos/pillow/Data.ts
// Original file: src/protos/pillow.proto


export interface Data {
  '_id'?: (string);
  'snoringRange'?: (string);
  'respirationRate'?: (string);
  'bodyTemperature'?: (string);
  'limbMovement'?: (string);
  'bloodOxygen'?: (string);
  'rem'?: (string);
  'hoursSleeping'?: (string);
  'heartRate'?: (string);
  'stresState'?: (string);
}

export interface Data__Output {
  '_id': (string);
  'snoringRange': (string);
  'respirationRate': (string);
  'bodyTemperature': (string);
  'limbMovement': (string);
  'bloodOxygen': (string);
  'rem': (string);
  'hoursSleeping': (string);
  'heartRate': (string);
  'stresState': (string);
}